Abstract searching of old TIs into a standalone component
